IGNITE-17132 wip

+++ modules/storage-rocksdb/src/main/java/org/apache/ignite/internal/storage/rocksdb/RocksDbMvPartitionStorage.java
@@ -27,6 +27,8 @@ import static org.apache.ignite.internal.util.ByteUtils.bytesToUuid;
 import static org.apache.ignite.internal.util.ByteUtils.putUuidToBytes;
 import static org.rocksdb.ReadTier.PERSISTED_TIER;
 
+import java.lang.invoke.MethodHandles;
+import java.lang.invoke.VarHandle;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 import java.nio.charset.StandardCharsets;
@@ -44,6 +46,7 @@ import org.apache.ignite.internal.storage.MvPartitionStorage;
 import org.apache.ignite.internal.storage.PartitionTimestampCursor;
 import org.apache.ignite.internal.storage.ReadResult;
 import org.apache.ignite.internal.storage.RowId;
+import org.apache.ignite.internal.storage.StorageClosedException;
 import org.apache.ignite.internal.storage.StorageException;
 import org.apache.ignite.internal.storage.TxIdMismatchException;
 import org.apache.ignite.internal.util.ByteUtils;
@@ -141,6 +144,16 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
             () -> ByteBuffer.allocate(MAX_KEY_SIZE).order(KEY_BYTE_ORDER)
     );
 
+    private static final VarHandle STARTED;
+
+    static {
+        try {
+            STARTED = MethodHandles.lookup().findVarHandle(RocksDbMvPartitionStorage.class, "started", boolean.class);
+        } catch (ReflectiveOperationException e) {
+            throw new ExceptionInInitializerError(e);
+        }
+    }
+
     /** Table storage instance. */
     private final RocksDbTableStorage tableStorage;
 
@@ -186,6 +199,8 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
     /** The value of {@link #lastAppliedIndex} persisted to the device at this moment. */
     private volatile long persistedIndex;
 
+    private volatile boolean started = true;
+
     /**
      * Constructor.
      *
@@ -210,9 +225,10 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
         persistedIndex = lastAppliedIndex;
     }
 
-    /** {@inheritDoc} */
     @Override
     public <V> V runConsistently(WriteClosure<V> closure) throws StorageException {
+        checkClosed();
+
         if (WRITE_BATCH.get() != null) {
             return closure.execute();
         } else {
@@ -238,9 +254,10 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
         }
     }
 
-    /** {@inheritDoc} */
     @Override
     public CompletableFuture<Void> flush() {
+        checkClosed();
+
         return tableStorage.awaitFlush(true);
     }
 
@@ -259,15 +276,17 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
         return requireWriteBatch();
     }
 
-    /** {@inheritDoc} */
     @Override
     public long lastAppliedIndex() {
+        checkClosed();
+
         return WRITE_BATCH.get() == null ? lastAppliedIndex : pendingAppliedIndex;
     }
 
-    /** {@inheritDoc} */
     @Override
     public void lastAppliedIndex(long lastAppliedIndex) throws StorageException {
+        checkClosed();
+
         WriteBatchWithIndex writeBatch = requireWriteBatch();
 
         try {
@@ -279,9 +298,10 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
         }
     }
 
-    /** {@inheritDoc} */
     @Override
     public long persistedIndex() {
+        checkClosed();
+
         return persistedIndex;
     }
 
@@ -313,10 +333,11 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
         return appliedIndexBytes == null ? 0 : ByteUtils.bytesToLong(appliedIndexBytes);
     }
 
-    /** {@inheritDoc} */
     @Override
     public @Nullable BinaryRow addWrite(RowId rowId, @Nullable BinaryRow row, UUID txId, UUID commitTableId, int commitPartitionId)
             throws TxIdMismatchException, StorageException {
+        checkClosed();
+
         @SuppressWarnings("resource") WriteBatchWithIndex writeBatch = requireWriteBatch();
 
         ByteBuffer keyBuf = prepareHeapKeyBuf(rowId);
@@ -394,9 +415,10 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
         return row.bytes();
     }
 
-    /** {@inheritDoc} */
     @Override
     public @Nullable BinaryRow abortWrite(RowId rowId) throws StorageException {
+        checkClosed();
+
         WriteBatchWithIndex writeBatch = requireWriteBatch();
 
         ByteBuffer keyBuf = prepareHeapKeyBuf(rowId);
@@ -420,9 +442,10 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
         }
     }
 
-    /** {@inheritDoc} */
     @Override
     public void commitWrite(RowId rowId, HybridTimestamp timestamp) throws StorageException {
+        checkClosed();
+
         WriteBatchWithIndex writeBatch = requireWriteBatch();
 
         ByteBuffer keyBuf = prepareHeapKeyBuf(rowId);
@@ -452,6 +475,8 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
 
     @Override
     public void addWriteCommitted(RowId rowId, BinaryRow row, HybridTimestamp commitTimestamp) throws StorageException {
+        checkClosed();
+
         @SuppressWarnings("resource") WriteBatchWithIndex writeBatch = requireWriteBatch();
 
         ByteBuffer keyBuf = prepareHeapKeyBuf(rowId);
@@ -467,9 +492,10 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
         }
     }
 
-    /** {@inheritDoc} */
     @Override
     public ReadResult read(RowId rowId, HybridTimestamp timestamp) throws StorageException {
+        checkClosed();
+
         if (rowId.partitionId() != partitionId) {
             throw new IllegalArgumentException(
                     String.format("RowId partition [%d] is not equal to storage partition [%d].", rowId.partitionId(), partitionId));
@@ -684,6 +710,8 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
 
     @Override
     public Cursor<ReadResult> scanVersions(RowId rowId) throws StorageException {
+        checkClosed();
+
         ByteBuffer keyBuf = prepareHeapKeyBuf(rowId);
 
         byte[] lowerBound = copyOf(keyBuf.array(), ROW_PREFIX_SIZE);
@@ -699,6 +727,13 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
         it.seek(lowerBound);
 
         return new RocksIteratorAdapter<>(it) {
+            @Override
+            public boolean hasNext() {
+                checkClosed();
+
+                return super.hasNext();
+            }
+
             @Override
             protected ReadResult decodeEntry(byte[] key, byte[] value) {
                 int keyLength = key.length;
@@ -718,10 +753,9 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
     }
 
     // TODO: IGNITE-16914 Play with prefix settings and benchmark results.
-    /** {@inheritDoc} */
     @Override
     public PartitionTimestampCursor scan(HybridTimestamp timestamp) throws StorageException {
-        Objects.requireNonNull(timestamp, "timestamp is null");
+        checkClosed();
 
         if (lookingForLatestVersions(timestamp)) {
             return new ScanLatestVersionsCursor();
@@ -743,6 +777,8 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
 
     @Override
     public @Nullable RowId closestRowId(RowId lowerBound) throws StorageException {
+        checkClosed();
+
         ByteBuffer keyBuf = prepareHeapKeyBuf(lowerBound).position(0).limit(ROW_PREFIX_SIZE);
 
         try (RocksIterator it = db.newIterator(cf, scanReadOptions)) {
@@ -798,6 +834,8 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
 
     @Override
     public long rowsCount() {
+        checkClosed();
+
         try (
                 var upperBound = new Slice(partitionEndPrefix());
                 var options = new ReadOptions().setIterateUpperBound(upperBound);
@@ -808,6 +846,8 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
             long size = 0;
 
             while (it.isValid()) {
+                checkClosed();
+
                 ++size;
                 it.next();
             }
@@ -817,27 +857,38 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
     }
 
     /**
-     * Deletes partition data from the storage.
+     * Closes the partition storage and removes data from it.
      */
     public void destroy() {
-        try (WriteBatch writeBatch = new WriteBatch()) {
-            writeBatch.delete(meta, lastAppliedIndexKey);
-
-            writeBatch.delete(meta, RocksDbMetaStorage.partitionIdKey(partitionId));
-
-            writeBatch.deleteRange(cf, partitionStartPrefix(), partitionEndPrefix());
-
-            db.write(writeOpts, writeBatch);
-        } catch (RocksDBException e) {
-            TableConfiguration tableCfg = tableStorage.configuration();
-
-            throw new StorageException("Failed to destroy partition " + partitionId + " of table " + tableCfg.name(), e);
-        }
+        close(true);
     }
 
-    /** {@inheritDoc} */
     @Override
     public void close() {
+        close(false);
+    }
+
+    private void close(boolean destroy) {
+        if (!STARTED.compareAndSet(this, true, false)) {
+            return;
+        }
+
+        if (destroy) {
+            try (WriteBatch writeBatch = new WriteBatch()) {
+                writeBatch.delete(meta, lastAppliedIndexKey);
+
+                writeBatch.delete(meta, RocksDbMetaStorage.partitionIdKey(partitionId));
+
+                writeBatch.deleteRange(cf, partitionStartPrefix(), partitionEndPrefix());
+
+                db.write(writeOpts, writeBatch);
+            } catch (RocksDBException e) {
+                TableConfiguration tableCfg = tableStorage.configuration();
+
+                throw new StorageException("Failed to destroy partition " + partitionId + " of table " + tableCfg.name(), e);
+            }
+        }
+
         RocksUtils.closeAll(persistedTierReadOpts, readOpts, writeOpts, scanReadOptions, upperBound);
     }
 
@@ -1070,6 +1121,8 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
     private final class ScanLatestVersionsCursor extends BasePartitionTimestampCursor {
         @Override
         public boolean hasNext() {
+            checkClosed();
+
             // Fast-path for consecutive invocations.
             if (next != null) {
                 return true;
@@ -1086,6 +1139,8 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
             ByteBuffer currentKeyBuffer = MV_KEY_BUFFER.get().position(0);
 
             while (true) {
+                checkClosed();
+
                 currentKeyBuffer.position(0);
 
                 // At this point, seekKeyBuf should contain row id that's above the one we already scanned, but not greater than any
@@ -1187,6 +1242,8 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
 
         @Override
         public boolean hasNext() {
+            checkClosed();
+
             // Fast-path for consecutive invocations.
             if (next != null) {
                 return true;
@@ -1203,6 +1260,8 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
             ByteBuffer directBuffer = MV_KEY_BUFFER.get().position(0);
 
             while (true) {
+                checkClosed();
+
                 it.seek(seekKeyBuf.array());
 
                 if (invalid(it)) {
@@ -1233,6 +1292,12 @@ public class RocksDbMvPartitionStorage implements MvPartitionStorage {
 
                 return true;
             }
+        }
+    }
+
+    private void checkClosed() {
+        if (!started) {
+            throw new StorageClosedException("Storage is already closed");
         }
     }
 }
